chore(suite): extract separate AddButton reused in AddCoinJoinAccountButton
